Introduce read task
